Engine can now import arbitrary problems.
Problems are now imported based on the problemID field provided in the POST request. Used importlib to dynamically import the correct problem module. This will only work if the problemID meta tag on the problem HTML pages matches the problem module name, so we might as well keep doing that.

Also moved the test case generation code from the individual problem module to the engine as that particular piece of code will always be the same no matter the module.