Code now runs in /root.
So apparently copying files to /tmp/ immediately after container creation
is a bad idea as files in /tmp/ can get deleted during bootup (which was
happening to files we copied over there). Tried to do everything in
/lovelace/ but then you have to "lxc exec ... -- mkdir /lovelace/" which
added another command that had to be executed, and it timed out for some
reason even though /lovelace/ was created. So now I'm running in /root/
for now, where copying files and running python3 works, and we get the
output pickle file. But for some reason it times out after that...

See: https://github.com/lxc/lxd/issues/2668#issuecomment-263755926