Modify code runner Dockerfile to support ARM architecture
Most of the code runner dockerfile was platform agnostic, but the Julia
installation was using a binary for x86_64 specifically.

I first re-wrote the docker file to use `$(uname -m)` to check the
platform, but the quoting and escaping in the RUN commands were a big
mess. Luckily there's a python tool called Jill which handles installing
Julia from the command line:
https://github.com/abelsiqueira/jill